requirements.txt
Django
asgiref
pytz
sqlparse
python-decouple
djangorestframework
gunicorn
redis
psycopg2
psycopg2-binary
requests
markdown
django-filter
django-cors-headers
